+++ src/renderer/src/runtime/runtime-git-client.ts
@@ -20,6 +20,7 @@ import { getCommitMessageModelDiscoveryHostKeyForScope } from '../../../shared/c
 import type { GitHistoryOptions, GitHistoryResult } from '../../../shared/git-history'
 import { getRepoIdFromWorktreeId } from '../../../shared/worktree-id'
 import { callRuntimeRpc, getActiveRuntimeTarget } from './runtime-rpc-client'
+import { toRuntimeWorktreeSelector } from './runtime-worktree-selector'
 
 export type RuntimeGenerateCommitMessageResult =
   | { success: true; message: string; agentLabel?: string }
@@ -122,7 +123,7 @@ export async function getRuntimeGitStatus(
   return callRuntimeRpc<GitStatusResult>(
     target,
     'git.status',
-    { worktree: context.worktreeId, ...includeIgnoredArgs },
+    { worktree: toRuntimeWorktreeSelector(context.worktreeId), ...includeIgnoredArgs },
     { timeoutMs: 15_000 }
   )
 }
@@ -145,7 +146,7 @@ export async function getRuntimeGitIgnoredPaths(
   return callRuntimeRpc<string[]>(
     target,
     'git.checkIgnored',
-    { worktree: context.worktreeId, paths },
+    { worktree: toRuntimeWorktreeSelector(context.worktreeId), paths },
     { timeoutMs: 15_000 }
   )
 }
@@ -165,7 +166,7 @@ export async function getRuntimeGitHistory(
   return callRuntimeRpc<GitHistoryResult>(
     target,
     'git.history',
-    { worktree: context.worktreeId, ...options },
+    { worktree: toRuntimeWorktreeSelector(context.worktreeId), ...options },
     { timeoutMs: 15_000 }
   )
 }
@@ -183,7 +184,7 @@ export async function getRuntimeGitConflictOperation(
   return callRuntimeRpc<GitConflictOperation>(
     target,
     'git.conflictOperation',
-    { worktree: context.worktreeId },
+    { worktree: toRuntimeWorktreeSelector(context.worktreeId) },
     { timeoutMs: 15_000 }
   )
 }
@@ -200,7 +201,7 @@ export async function abortRuntimeGitMerge(context: RuntimeGitContext): Promise<
   await callRuntimeRpc(
     target,
     'git.abortMerge',
-    { worktree: context.worktreeId },
+    { worktree: toRuntimeWorktreeSelector(context.worktreeId) },
     { timeoutMs: 30_000 }
   )
 }
@@ -217,7 +218,7 @@ export async function abortRuntimeGitRebase(context: RuntimeGitContext): Promise
   await callRuntimeRpc(
     target,
     'git.abortRebase',
-    { worktree: context.worktreeId },
+    { worktree: toRuntimeWorktreeSelector(context.worktreeId) },
     { timeoutMs: 30_000 }
   )
 }
@@ -239,7 +240,7 @@ export async function getRuntimeGitDiff(
   return callRuntimeRpc<GitDiffResult>(
     target,
     'git.diff',
-    { worktree: context.worktreeId, ...args },
+    { worktree: toRuntimeWorktreeSelector(context.worktreeId), ...args },
     { timeoutMs: 15_000 }
   )
 }
@@ -259,7 +260,7 @@ export async function getRuntimeGitBranchCompare(
   return callRuntimeRpc<GitBranchCompareResult>(
     target,
     'git.branchCompare',
-    { worktree: context.worktreeId, baseRef },
+    { worktree: toRuntimeWorktreeSelector(context.worktreeId), baseRef },
     { timeoutMs: 15_000 }
   )
 }
@@ -279,7 +280,7 @@ export async function getRuntimeGitCommitCompare(
   return callRuntimeRpc<GitCommitCompareResult>(
     target,
     'git.commitCompare',
-    { worktree: context.worktreeId, commitId },
+    { worktree: toRuntimeWorktreeSelector(context.worktreeId), commitId },
     { timeoutMs: 15_000 }
   )
 }
@@ -299,7 +300,10 @@ export async function getRuntimeGitUpstreamStatus(
   return callRuntimeRpc<GitUpstreamStatus>(
     target,
     'git.upstreamStatus',
-    { worktree: context.worktreeId, ...(pushTarget ? { pushTarget } : {}) },
+    {
+      worktree: toRuntimeWorktreeSelector(context.worktreeId),
+      ...(pushTarget ? { pushTarget } : {})
+    },
     { timeoutMs: 15_000 }
   )
 }
@@ -320,7 +324,10 @@ export async function fetchRuntimeGit(
   await callRuntimeRpc(
     target,
     'git.fetch',
-    { worktree: context.worktreeId, ...(pushTarget ? { pushTarget } : {}) },
+    {
+      worktree: toRuntimeWorktreeSelector(context.worktreeId),
+      ...(pushTarget ? { pushTarget } : {})
+    },
     { timeoutMs: 30_000 }
   )
 }
@@ -341,7 +348,10 @@ export async function pullRuntimeGit(
   await callRuntimeRpc(
     target,
     'git.pull',
-    { worktree: context.worktreeId, ...(pushTarget ? { pushTarget } : {}) },
+    {
+      worktree: toRuntimeWorktreeSelector(context.worktreeId),
+      ...(pushTarget ? { pushTarget } : {})
+    },
     { timeoutMs: 30_000 }
   )
 }
@@ -362,7 +372,10 @@ export async function fastForwardRuntimeGit(
   await callRuntimeRpc(
     target,
     'git.fastForward',
-    { worktree: context.worktreeId, ...(pushTarget ? { pushTarget } : {}) },
+    {
+      worktree: toRuntimeWorktreeSelector(context.worktreeId),
+      ...(pushTarget ? { pushTarget } : {})
+    },
     { timeoutMs: 30_000 }
   )
 }
@@ -383,7 +396,7 @@ export async function rebaseRuntimeGitFromBase(
   await callRuntimeRpc(
     target,
     'git.rebaseFromBase',
-    { worktree: context.worktreeId, baseRef },
+    { worktree: toRuntimeWorktreeSelector(context.worktreeId), baseRef },
     { timeoutMs: 30_000 }
   )
 }
@@ -407,7 +420,7 @@ export async function pushRuntimeGit(
     target,
     'git.push',
     {
-      worktree: context.worktreeId,
+      worktree: toRuntimeWorktreeSelector(context.worktreeId),
       ...(args.publish !== undefined ? { publish: args.publish } : {}),
       ...(args.pushTarget !== undefined ? { pushTarget: args.pushTarget } : {}),
       ...(args.forceWithLease !== undefined ? { forceWithLease: args.forceWithLease } : {})
@@ -437,7 +450,7 @@ export async function getRuntimeGitBranchDiff(
   return callRuntimeRpc<GitDiffResult>(
     target,
     'git.branchDiff',
-    { worktree: context.worktreeId, ...args },
+    { worktree: toRuntimeWorktreeSelector(context.worktreeId), ...args },
     { timeoutMs: 15_000 }
   )
 }
@@ -465,7 +478,7 @@ export async function getRuntimeGitCommitDiff(
   return callRuntimeRpc<GitDiffResult>(
     target,
     'git.commitDiff',
-    { worktree: context.worktreeId, ...args },
+    { worktree: toRuntimeWorktreeSelector(context.worktreeId), ...args },
     { timeoutMs: 15_000 }
   )
 }
@@ -485,7 +498,7 @@ export async function commitRuntimeGit(
   return callRuntimeRpc<{ success: boolean; error?: string }>(
     target,
     'git.commit',
-    { worktree: context.worktreeId, message },
+    { worktree: toRuntimeWorktreeSelector(context.worktreeId), message },
     { timeoutMs: 30_000 }
   )
 }
@@ -511,7 +524,7 @@ export async function generateRuntimeCommitMessage(
     target,
     'git.generateCommitMessage',
     {
-      worktree: context.worktreeId,
+      worktree: toRuntimeWorktreeSelector(context.worktreeId),
       ...getRuntimeCommitMessageSettings(context.settings, context.connectionId),
       ...(overrides?.sourceControlAiResolvedParams
         ? { sourceControlAiResolvedParams: overrides.sourceControlAiResolvedParams }
@@ -539,7 +552,7 @@ export async function discoverRuntimeCommitMessageModels(
     target,
     'git.discoverCommitMessageModels',
     {
-      worktree: context.worktreeId,
+      worktree: toRuntimeWorktreeSelector(context.worktreeId),
       agentId,
       ...(context.settings?.agentCmdOverrides
         ? { agentCmdOverrides: context.settings.agentCmdOverrides }
@@ -563,7 +576,7 @@ export async function cancelRuntimeGenerateCommitMessage(
   await callRuntimeRpc(
     target,
     'git.cancelGenerateCommitMessage',
-    { worktree: context.worktreeId },
+    { worktree: toRuntimeWorktreeSelector(context.worktreeId) },
     { timeoutMs: 5_000 }
   )
 }
@@ -591,7 +604,7 @@ export async function generateRuntimePullRequestFields(
     target,
     'git.generatePullRequestFields',
     {
-      worktree: context.worktreeId,
+      worktree: toRuntimeWorktreeSelector(context.worktreeId),
       ...input,
       ...getRuntimeCommitMessageSettings(context.settings, context.connectionId),
       ...(overrides?.sourceControlAiResolvedParams
@@ -618,7 +631,7 @@ export async function cancelRuntimeGeneratePullRequestFields(
   await callRuntimeRpc(
     target,
     'git.cancelGeneratePullRequestFields',
-    { worktree: context.worktreeId },
+    { worktree: toRuntimeWorktreeSelector(context.worktreeId) },
     { timeoutMs: 5_000 }
   )
 }
@@ -639,7 +652,7 @@ export async function stageRuntimeGitPath(
   await callRuntimeRpc(
     target,
     'git.stage',
-    { worktree: context.worktreeId, filePath },
+    { worktree: toRuntimeWorktreeSelector(context.worktreeId), filePath },
     { timeoutMs: 15_000 }
   )
 }
@@ -660,7 +673,7 @@ export async function bulkStageRuntimeGitPaths(
   await callRuntimeRpc(
     target,
     'git.bulkStage',
-    { worktree: context.worktreeId, filePaths },
+    { worktree: toRuntimeWorktreeSelector(context.worktreeId), filePaths },
     { timeoutMs: 15_000 }
   )
 }
@@ -681,7 +694,7 @@ export async function unstageRuntimeGitPath(
   await callRuntimeRpc(
     target,
     'git.unstage',
-    { worktree: context.worktreeId, filePath },
+    { worktree: toRuntimeWorktreeSelector(context.worktreeId), filePath },
     { timeoutMs: 15_000 }
   )
 }
@@ -702,7 +715,7 @@ export async function bulkUnstageRuntimeGitPaths(
   await callRuntimeRpc(
     target,
     'git.bulkUnstage',
-    { worktree: context.worktreeId, filePaths },
+    { worktree: toRuntimeWorktreeSelector(context.worktreeId), filePaths },
     { timeoutMs: 15_000 }
   )
 }
@@ -723,7 +736,7 @@ export async function bulkDiscardRuntimeGitPaths(
   await callRuntimeRpc(
     target,
     'git.bulkDiscard',
-    { worktree: context.worktreeId, filePaths },
+    { worktree: toRuntimeWorktreeSelector(context.worktreeId), filePaths },
     { timeoutMs: 15_000 }
   )
 }
@@ -744,7 +757,7 @@ export async function discardRuntimeGitPath(
   await callRuntimeRpc(
     target,
     'git.discard',
-    { worktree: context.worktreeId, filePath },
+    { worktree: toRuntimeWorktreeSelector(context.worktreeId), filePath },
     { timeoutMs: 15_000 }
   )
 }
@@ -765,7 +778,11 @@ export async function getRuntimeGitRemoteFileUrl(
   return callRuntimeRpc<string | null>(
     target,
     'git.remoteFileUrl',
-    { worktree: context.worktreeId, relativePath: args.relativePath, line: args.line },
+    {
+      worktree: toRuntimeWorktreeSelector(context.worktreeId),
+      relativePath: args.relativePath,
+      line: args.line
+    },
     { timeoutMs: 15_000 }
   )
 }
